blog
During this time period Russia was undergoing a revolution. The socialist party was at the head of all of this. A couple days before this newspaper was published there was a stand off between Jews and the Czars. The police opened fire on the crowd and blood ran through the streets. Im curious if there was any socialist pressure in Alexandria? How would the people reading this article react to such a thing? Would they get ideas of their own or just think nothing of it. Newspapers are a marvelous median to display information over. Back in 1905 information like that was able to travel all around the world in just a matter of a few days. These attacks on socialist groups ( mainly comprised of jews) were prevalent throughout this entire time period. I am also curious if there was any prejudice acts of hate that emulate the ones happening in Russia. http://www.jewishvirtuallibrary.org/jsource/History/pogroms.html