Blog Post and Image
This includes 3 links to outside sources, but I didn't know how to make them into hyperlinks in markdown so theres just links in the middle of the text. Also the syllabus says to include links to at least 2 related past blog posts, but there was only one that was somewhat similar, so i linked that one.